requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography
prometheus-client==0.19.0
slowapi==0.1.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psutil==5.9.6